RC 0.1

+++ source/tinycmd/tinycmdapi.h
@@ -14,7 +14,7 @@ uint8_t tinycmd_rgb_all(uint8_t on, uint8_t r, uint8_t g, uint8_t b, uint8_t rsp
 uint8_t tinycmd_rgb_pos(uint8_t pos, uint8_t r, uint8_t g, uint8_t b, uint8_t rsp);
 uint8_t tinycmd_rgb_range(uint8_t num, uint8_t offset, uint8_t r, uint8_t g, uint8_t b, uint8_t rsp);
 uint8_t tinycmd_rgb_buffer(uint8_t num, uint8_t offset, uint8_t *led, uint8_t rsp);
-uint8_t tinycmd_rgb_effect_speed(uint8_t speed, uint8_t rsp);
+uint8_t tinycmd_rgb_effect_speed(uint16_t speed, uint8_t rsp);
 uint8_t tinycmd_rgb_effect_on(uint8_t on, uint8_t rsp);
 
 uint8_t tinycmd_rgb_set_effect(uint8_t index, uint8_t rsp);
